# api-contracts.yaml
openapi: 3.0.3
info:
  title: B2B Platform API
  version: 1.0.0
  description: >
    B2B procurement platform API: requests, keys, suppliers, moderation, email sending,
    inbound reply tracking. Attachments are uploaded separately and referenced by ID.
servers:
  - url: http://localhost:8000/api/v1
    description: Local
  - url: https://{host}/api/v1
    description: Production
    variables:
      host:
        default: app.example.com
...
security:
  - bearerAuth: []
paths:
...
  /auth/oauth/google/callback:
    get:
      tags: [Auth]
      summary: OAuth callback (Google)
      security: []
      parameters:
        - in: query
          name: code
          required: true
          schema: { type: string }
      responses:
        "200":
          description: JWT issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
...
  /auth/policy:
    put:
      tags: [Auth]
      summary: Update email deletion policy
      description: >
        append_only = user cannot delete messages in UI (system remains append-only and auditable).
        allow_delete = user can hide (soft-delete) messages for their view.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateEmailPolicyRequest" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }

  # ----------------------------
  # REQUESTS
  # ----------------------------
...
  /user/attachments:
    post:
      tags: [Attachments]
      summary: Upload an attachment (for email sending)
      description: >
        Uploads a file and returns attachment_id. Later used in /send endpoints.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                  description: Optional display name
      responses:
        "200":
          description: Attachment uploaded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attachment" }
        "413":
          description: Attachment too large
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    get:
      tags: [Attachments]
      summary: List user attachments
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - in: query
          name: offset
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: Attachments
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttachmentListResponse" }
  /user/attachments/{attachmentId}:
    get:
      tags: [Attachments]
      summary: Get attachment metadata
      parameters:
        - $ref: "#/components/parameters/AttachmentId"
      responses:
        "200":
          description: Attachment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attachment" }

    delete:
      tags: [Attachments]
      summary: Delete attachment (soft delete)
      parameters:
        - $ref: "#/components/parameters/AttachmentId"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenericOkResponse" }
  /user/attachments/{attachmentId}/download:
    get:
      tags: [Attachments]
      summary: Download attachment file
      parameters:
        - $ref: "#/components/parameters/AttachmentId"
      responses:
        "200":
          description: File stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

  # ----------------------------
  # USER MESSAGING (send + replies)
  # ----------------------------
...
  /moderator/tasks:
    get:
      tags: [ModeratorTasks]
      summary: List moderation tasks
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, parsed, needs_manual, done]
        - in: query
          name: limit
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModeratorTaskListResponse" }
  /moderator/tasks/{taskId}:
    get:
      tags: [ModeratorTasks]
      summary: Get moderation task detail
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModeratorTaskDetail" }
  /moderator/requests/{requestId}/start-parsing:
    post:
      tags: [ModeratorTasks]
      summary: Start parsing URLs for a submitted request
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Parsing started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenericOkResponse" }

  # ----------------------------
  # MODERATOR SUPPLIERS (database view + create)
  # ----------------------------
  /moderator/suppliers:
    get:
      tags: [ModeratorSuppliers]
      summary: Browse/search suppliers database (moderator view)
      parameters:
        - in: query
          name: q
          schema: { type: string, description: "Search by name/INN/domain/key" }
        - in: query
          name: limit
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - in: query
          name: offset
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: Suppliers
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModeratorSupplierListResponse" }

    post:
      tags: [ModeratorSuppliers]
      summary: Add supplier to database (moderated)
      description: >
        Required fields: url_original, name, inn. After creation, enrich via Checko API.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSupplierRequest" }
      responses:
        "200":
          description: Supplier created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Supplier" }
  /moderator/suppliers/{supplierId}:
    get:
      tags: [ModeratorSuppliers]
      summary: Get supplier detail (with Checko data)
      parameters:
        - $ref: "#/components/parameters/SupplierId"
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Supplier" }

    put:
      tags: [ModeratorSuppliers]
      summary: Update supplier (e.g., refresh Checko data)
      parameters:
        - $ref: "#/components/parameters/SupplierId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSupplierRequest" }
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Supplier" }

  # ----------------------------
  # BLACKLIST
  # ----------------------------
  /moderator/blacklist/domains:
    post:
      tags: [Blacklist]
      summary: Add domain to blacklist (ban entire domain)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddBlacklistDomainRequest" }
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenericOkResponse" }

    get:
      tags: [Blacklist]
      summary: List blacklisted domains
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 200, minimum: 1, maximum: 500 }
      responses:
        "200":
          description: Blacklist
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlacklistDomainListResponse" }
  /moderator/blacklist/domains/{domain}:
    delete:
      tags: [Blacklist]
      summary: Remove domain from blacklist (unban)
      parameters:
        - in: path
          name: domain
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenericOkResponse" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RequestId:
      in: path
      name: requestId
      required: true
      schema: { type: integer, minimum: 1 }
    TaskId:
      in: path
      name: taskId
      required: true
      schema: { type: integer, minimum: 1 }
    SupplierId:
      in: path
      name: supplierId
      required: true
      schema: { type: integer, minimum: 1 }
    AttachmentId:
      in: path
      name: attachmentId
      required: true
      schema: { type: integer, minimum: 1 }
...